Add LaTeX abbreviations for subscripts, superscripts and
operators

* lisp/dynexp-abbrev.el: Add new abbreviations for mathematical
notation.  Add "lsh" and "lhh" for subscript h with braces.  Add "hhh"
and "hsh" for superscript h with braces.  Add "sop" for \Opp operator
and "soph" for \Opp with subscript h.  Add "sh" and "dsh" for \h with
conditional and forced dollar wrapping.  Add "ssup" as alternative
abbreviation for \sup.